separates list component for reuse
